Add factor out components from outbound shipment ListView
