+++ packages/invoices/src/OutboundShipment/ListView/ListView.tsx
@@ -1,77 +1,25 @@
-import React, { FC, useEffect, useRef, useState } from 'react';
+import React, { FC, useState } from 'react';
 import { useNavigate } from 'react-router';
 
 import {
-  DownloadIcon,
-  PlusCircleIcon,
-  PrinterIcon,
   DataTable,
   useColumns,
-  useNotification,
   InvoiceRow,
-  DropdownMenu,
-  DropdownMenuItem,
-  AppBarContentPortal,
-  useTranslation,
   useListData,
   getNameAndColorColumn,
-  DeleteIcon,
   TableProvider,
   createTableStore,
-  useTableStore,
   Color,
-  AppBarButtonsPortal,
-  BookIcon,
-  ButtonWithIcon,
-  Grid,
   OutboundShipmentStatus,
 } from '@openmsupply-client/common';
 import { OutboundShipmentListViewApi } from './api';
-import { ExternalURL } from '@openmsupply-client/config';
+
 import { CustomerSearch } from './CustomerSearch';
 import { getStatusTranslation } from '../utils';
-
-const ListViewToolBar: FC<{
-  onDelete: (toDelete: InvoiceRow[]) => void;
-  data?: InvoiceRow[];
-}> = ({ onDelete, data }) => {
-  const t = useTranslation();
-
-  const { success, info } = useNotification();
-
-  const { selectedRows } = useTableStore(state => ({
-    selectedRows: Object.keys(state.rowState)
-      .filter(id => state.rowState[id]?.isSelected)
-      .map(selectedId => data?.find(({ id }) => selectedId === id))
-      .filter(Boolean) as InvoiceRow[],
-  }));
-
-  const deleteAction = () => {
-    if (selectedRows && selectedRows?.length > 0) {
-      onDelete(selectedRows);
-      success(`Deleted ${selectedRows?.length} invoices`)();
-    } else {
-      info('Select rows to delete them')();
-    }
-  };
-
-  const ref = useRef(deleteAction);
-
-  useEffect(() => {
-    ref.current = deleteAction;
-  }, [selectedRows]);
-
-  return (
-    <DropdownMenu label="Select">
-      <DropdownMenuItem IconComponent={DeleteIcon} onClick={deleteAction}>
-        {t('button.delete-lines')}
-      </DropdownMenuItem>
-    </DropdownMenu>
-  );
-};
+import { Toolbar } from './Toolbar';
+import { AppBarButtons } from './AppBarButtons';
 
 export const OutboundShipmentListViewComponent: FC = () => {
-  const { info, success } = useNotification();
   const navigate = useNavigate();
 
   const {
@@ -138,34 +86,9 @@ export const OutboundShipmentListViewComponent: FC = () => {
           createInvoice();
         }}
       />
-      <AppBarContentPortal sx={{ paddingBottom: '16px' }}>
-        <ListViewToolBar onDelete={onDelete} data={data} />
-      </AppBarContentPortal>
 
-      <AppBarButtonsPortal>
-        <Grid container gap={1}>
-          <ButtonWithIcon
-            Icon={<PlusCircleIcon />}
-            labelKey="button.new-shipment"
-            onClick={() => setOpen(true)}
-          />
-          <ButtonWithIcon
-            Icon={<DownloadIcon />}
-            labelKey="button.export"
-            onClick={success('Downloaded successfully')}
-          />
-          <ButtonWithIcon
-            Icon={<PrinterIcon />}
-            labelKey="button.print"
-            onClick={info('No printer detected')}
-          />
-          <ButtonWithIcon
-            Icon={<BookIcon />}
-            labelKey="button.docs"
-            onClick={() => (location.href = ExternalURL.PublicDocs)}
-          />
-        </Grid>
-      </AppBarButtonsPortal>
+      <Toolbar onDelete={onDelete} data={data} />
+      <AppBarButtons onCreate={setOpen} />
 
       <DataTable
         pagination={{ ...pagination, total: totalCount }}
